+++ src/common/lib/client/defaultrest.ts
@@ -1,6 +1,7 @@
 import { BaseRest } from './baserest';
 import ClientOptions from '../../types/ClientOptions';
 import { allCommonModules } from './modulesmap';
+import Platform from 'common/platform';
 
 /**
  `DefaultRest` is the class that the non tree-shakable version of the SDK exports as `Rest`. It ensures that this version of the SDK includes all of the functionality which is optionally available in the tree-shakable version.
@@ -8,5 +9,17 @@ import { allCommonModules } from './modulesmap';
 export class DefaultRest extends BaseRest {
   constructor(options: ClientOptions | string) {
     super(options, allCommonModules);
+  }
+
+  private static _Crypto: typeof Platform.Crypto = null;
+  static get Crypto() {
+    if (this._Crypto === null) {
+      throw new Error('Encryption not enabled; use ably.encryption.js instead');
+    }
+
+    return this._Crypto;
+  }
+  static set Crypto(newValue: typeof Platform.Crypto) {
+    this._Crypto = newValue;
   }
 }
